Verify every fill by simulation before issuing it
The rasterizer comes first, because the fill check needs it: one brush, one
line walk, one flood fill, shared by the preview and the verifier, so a preview
that looks right and an execution that leaks cannot disagree.

A fill is accepted only when a flood of the simulated canvas, painted with the
outlines the plan will draw, stops inside its region. One pixel outside is a
refusal: a leak runs until it meets the next painted thing, so it is not a
matter of degree. Refused fills cost nothing, because what they would have
covered is still unpainted and the stroke planner picks it up.

Tests build outlines with one-pixel gaps on each of the four edges and assert
both halves: the refusal, and a canvas left untouched.

+++ src/penplan/model.py
@@ -226,7 +226,7 @@ class DrawPlan:
     width: int
     height: int
     palette: tuple[Rgb, ...]
-    brush_widths: tuple[int, ...]
+    brush_widths: tuple[float, ...]
     steps: tuple[Step, ...]
     report: PlanReport
